병합 충돌 해결 및 SeedTable의 ActiveTime int->float

+++ Assets/Scripts/DataTable/Table_Seed.cs
@@ -17,7 +17,7 @@ namespace DataTable
         {
             public int Index;
             public string Type;
-            public int ActiveTime;
+            public float ActiveTime;
             public string SpritePath;
         }
